apnedoctors_app/AI-Powered Symtom Checker/requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
python-dotenv==0.20.0
sqlalchemy==1.4.39
alembic==1.8.1
psycopg2-binary==2.9.3
pydantic==1.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytorch==1.11.0
transformers==4.20.1
numpy==1.23.1
pandas==1.4.3
aiohttp==3.8.1
pytest==7.1.2
python-multipart==0.0.5